requirements.txt
streamlit
markdown2
beautifulsoup4
weasyprint
python-docx